fix(ui): implement ResizeHandle ans Sidebar as component
